Testing - part 2, Dto

+++ src/test/java/com/example/smartwardrobe/controller/OutfitControllerTest.java
@@ -1,27 +1,20 @@
 package com.example.smartwardrobe.controller;
 
-import static org.mockito.Mockito.any;
-import static org.mockito.Mockito.doNothing;
-import static org.mockito.Mockito.when;
-
-import com.example.smartwardrobe.enums.CoatCategory;
-import com.example.smartwardrobe.model.Outfit;
 import com.example.smartwardrobe.service.OutfitService;
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-import java.util.ArrayList;
-
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.http.MediaType;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+
+import java.util.ArrayList;
+
+import static org.mockito.Mockito.*;
 
 @ContextConfiguration(classes = {OutfitController.class})
 @ExtendWith(SpringExtension.class)
@@ -55,7 +48,7 @@ class OutfitControllerTest {
 
     @Test
     void testDeleteOutfitById3() throws Exception {
-        doNothing().when(this.outfitService).deleteOutfitById((Long) any());
+        doNothing().when(this.outfitService).deleteOutfitById(any());
         MockHttpServletRequestBuilder requestBuilder = MockMvcRequestBuilders.delete("/outfit/{id}", 123L);
         MockMvcBuilders.standaloneSetup(this.outfitController)
                 .build()
@@ -99,46 +92,5 @@ class OutfitControllerTest {
                 .andExpect(MockMvcResultMatchers.content().string("[]"));
     }
 
-//    @Test
-//    void testGetOutfitById() throws Exception {
-//        Outfit outfit = new Outfit();
-//        outfit.setCoat(CoatCategory.JACKET);
-//        outfit.setDescription("The characteristics of someone or something");
-//        outfit.setHistories(new ArrayList<>());
-//        outfit.setId(123L);
-//        outfit.setItems(new ArrayList<>());
-//        when(this.outfitService.findOutfitById((Long) any())).thenReturn(outfit);
-//        MockHttpServletRequestBuilder requestBuilder = MockMvcRequestBuilders.get("/outfit/{id}", 123L);
-//        MockMvcBuilders.standaloneSetup(this.outfitController)
-//                .build()
-//                .perform(requestBuilder)
-//                .andExpect(MockMvcResultMatchers.status().isOk())
-//                .andExpect(MockMvcResultMatchers.content().contentType("application/json"))
-//                .andExpect(MockMvcResultMatchers.content()
-//                        .string(
-//                                "{\"id\":123,\"description\":\"The characteristics of someone or something\",\"coat\":\"JACKET\",\"items\":[]}"));
-//    }
-
-//    @Test
-//    void testSaveOutfit() throws Exception {
-//        when(this.outfitService.findAllOutfits()).thenReturn(new ArrayList<>());
-//
-//        Outfit outfit = new Outfit();
-//        outfit.setCoat(CoatCategory.JACKET);
-//        outfit.setDescription("The characteristics of someone or something");
-//        outfit.setHistories(new ArrayList<>());
-//        outfit.setId(123L);
-//        outfit.setItems(new ArrayList<>());
-//        String content = (new ObjectMapper()).writeValueAsString(outfit);
-//        MockHttpServletRequestBuilder requestBuilder = MockMvcRequestBuilders.get("/outfit")
-//                .contentType(MediaType.APPLICATION_JSON)
-//                .content(content);
-//        MockMvcBuilders.standaloneSetup(this.outfitController)
-//                .build()
-//                .perform(requestBuilder)
-//                .andExpect(MockMvcResultMatchers.status().isOk())
-//                .andExpect(MockMvcResultMatchers.content().contentType("application/json"))
-//                .andExpect(MockMvcResultMatchers.content().string("[]"));
-//    }
 }
